+++ Locales/esES.lua
@@ -1,4 +1,6 @@
-local L = LibStub("AceLocale-3.0"):NewLocale("HealthBarColor", "esES")
+local addonName = ...
+
+local L = LibStub("AceLocale-3.0"):NewLocale(addonName, "esES")
 if not L then return end
 
 --@localization(locale="esES", format="lua_additive_table")@