+++ Locales/ptBR.lua
@@ -1,4 +1,6 @@
-local L = LibStub("AceLocale-3.0"):NewLocale("HealthBarColor", "ptBR")
+local addonName = ...
+
+local L = LibStub("AceLocale-3.0"):NewLocale(addonName, "ptBR")
 if not L then return end
 
 --@localization(locale="ptBR", format="lua_additive_table")@