+++ Locales/ruRU.lua
@@ -1,4 +1,6 @@
-local L = LibStub("AceLocale-3.0"):NewLocale("HealthBarColor", "ruRU")
+local addonName = ...
+
+local L = LibStub("AceLocale-3.0"):NewLocale(addonName, "ruRU")
 if not L then return end
 
 --@localization(locale="ruRU", format="lua_additive_table")@